requirements.txt
anthropic
pydantic
pydantic-ai
langgraph
logfire
python-dotenv
fastapi
uvicorn
httpx
starlette
jinja2
